requirements.txt
openai
python-docx
pypdf
python-dotenv
langchain
langchain-openai
chromadb
fastapi
uvicorn